SB-672 add cert to truststore on repository creation
