+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/booters/StorageBooter.java
@@ -7,12 +7,9 @@ import org.carlspring.strongbox.providers.layout.LayoutProviderRegistry;
 import org.carlspring.strongbox.repository.RepositoryManagementStrategyException;
 import org.carlspring.strongbox.resource.ConfigurationResourceResolver;
 import org.carlspring.strongbox.services.RepositoryManagementService;
-import org.carlspring.strongbox.services.TrustStoreService;
-import org.carlspring.strongbox.services.support.TrustStoreCertificationAdditionException;
 import org.carlspring.strongbox.storage.Storage;
 import org.carlspring.strongbox.storage.repository.Repository;
 import org.carlspring.strongbox.storage.repository.RepositoryStatusEnum;
-import org.carlspring.strongbox.storage.repository.remote.RemoteRepository;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -46,9 +43,6 @@ public class StorageBooter
 
     @Inject
     private RepositoryManagementService repositoryManagementService;
-
-    @Inject
-    private TrustStoreService trustStoreService;
 
     private File lockFile = new File(ConfigurationResourceResolver.getVaultDirectory(), "storage-booter.lock");
 
@@ -229,27 +223,8 @@ public class StorageBooter
         {
             repositoryManagementService.putInService(storage.getId(), repositoryId);
         }
-
-        if (repository.getRemoteRepository() != null)
-        {
-            initializeRemoteRepository(repository.getRemoteRepository());
-        }
     }
 
-    private void initializeRemoteRepository(RemoteRepository remoteRepository)
-    {
-        if (remoteRepository.isAutoImportRemoteSSLCertificate())
-        {
-            try
-            {
-                trustStoreService.addSslCertificatesToTrustStore(remoteRepository.getUrl());
-            }
-            catch (IOException | TrustStoreCertificationAdditionException e)
-            {
-                logger.error("Could not import remote SSL certificate to trust store", e);
-            }
-        }
-    }
 
     public RepositoryManagementService getRepositoryManagementService()
     {
